refactor: Nav 컴포넌트 시멘틱태그 적용 및 이름 수정
1. 시멘틱태그 적용
	- 사이드바는 nav가 아닌 aside 태그를 사용한다는 점을 알게 되어, 이에 맞춰 태그 변경
2. 컴포넌트 이름 수정
	- 위와 같은 이유로 컴포넌트 이름 역시 변경
